requirements.txt
Django>=4.2.0
Pillow>=10.0.0
django-ckeditor>=6.7.0
django-crispy-forms>=2.0
crispy-bootstrap5>=0.7
django-allauth>=0.57.0
django-environ>=0.11.2
whitenoise>=6.6.0
gunicorn>=21.2.0
psycopg2-binary>=2.9.7
python-decouple>=3.8
